refactor: Move `updateBufferGraphic` to separate file
